Create ConnectionManager.java
Connection manager class is a class that would be used by the client to connect to the game server. For now the port number is fixed and both server and client apps would have to run on the same machine. The class is responsible only for connecting client to the server