Create MessageFormatter.java
MessageFormatter class is created to handle the creation of the message in the special format that the user will send to the server. Now it is implements the simplest message format which is all the necessary information combined in one string and separated by a single special character, in this case “:” colon. Possible improvement: design custom message format (xml, json, data structure )