Create DisplayManager.java
DisplayManager Class is responsible for displaying current state of the game. All the game session data is saved and, therefore, changed, on the server side. For the sake of simplicity of the communication between the client and the server this class gets current state of the game as a string with all the necessary information about the game state and simply prints it.