Create GameInitialiser.java
GameInitialiser class is a class that is responsible for preparing a player to the game. Currently, it simply gets the rules of the game  from the server , but if needed it can be expanded to do other preparation tasks like manage stakes, configure the game (e.g. choose a type of blackjack) etc.